+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  - Introduce a new detailed exception `\Wizaplace\SDK\Exception\ProductNotFound`
  - Introduce a new detailed exception `\Wizaplace\SDK\Exception\ReviewsAreDisabled`
  - Introduce a new detailed exception `\Wizaplace\SDK\Exception\SenderIsAlsoRecipient`
+ - Introduce a new detailed exception `\Wizaplace\SDK\Exception\CompanyHasNoAdministrator`
 
 ### Bugfixes